requirements.txt
django
pillow
wagtail
djangorestframework
